+++ packages/stream_feeds/lib/src/state/comment_reaction_list.dart
@@ -8,7 +8,7 @@ import '../models/feeds_reaction_data.dart';
 import '../models/query_configuration.dart';
 import '../repository/comments_repository.dart';
 import 'comment_reaction_list_state.dart';
-import 'event/comment_reaction_list_event_handler.dart';
+import 'event/handler/comment_reaction_list_event_handler.dart';
 import 'query/comment_reactions_query.dart';
 import 'state_notifier_extentions.dart';