+++ packages/stream_feeds/lib/src/state/poll_list.dart
@@ -7,7 +7,7 @@ import 'package:stream_core/stream_core.dart';
 import '../models/poll_data.dart';
 import '../models/query_configuration.dart';
 import '../repository/polls_repository.dart';
-import 'event/poll_list_event_handler.dart';
+import 'event/handler/poll_list_event_handler.dart';
 import 'poll_list_state.dart';
 import 'query/polls_query.dart';
 import 'state_notifier_extentions.dart';